Fase_5/Hackathon_fase5-main/requirements.txt
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pillow>=10.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
requests>=2.31.0
reportlab>=4.0.0
python-dateutil>=2.8.0
